requirements.txt
fastapi
uvicorn
streamlit
gradio
transformers
sentence-transformers
faiss-cpu
torch>=2.0
langchain
pandas
numpy
openpyxl
nltk
pytest
black
flake8
datasets
accelerate
wandb
accelerate
streamlit
torch
datasets
transformers
peft
huggingface_hub
wandb
openpyxl